Add source code for wind direction
The ADC values were measured in the lab by us. It does not correspond to
the values from the documentation, since the original values work with
a reference voltage of 5V, whereas we are using 3.3V with R10K in the
voltage divider.

We have also implemented only 8 possible direction values, since the
original resolution of 14 was too unreliable on the hardware level.